replace new config implementation with old one
