+++ src/config_old/setters.rs
@@ -1,5 +1,5 @@
 use super::Config;
-use crate::config::OVERRIDE_FILEPATH;
+use crate::config_old::OVERRIDE_FILEPATH;
 use std::path::PathBuf;
 use toml::{value::Map, Value};
