+++ tests/fixtures/config_1/config.toml
@@ -5,7 +5,7 @@ Test3 = '{{ echo "test_3" }}'
 Test4 = '{{ echo $TEST_VAR }}'
 Test5 = { value = "test_5", expose = true }
 
-[env]
+[envs]
 AWS_PROFILE = "default"
 AWS_DEFAULT_REGION = "eu-central-1"
