+++ tests/fixtures/config_1/template.yml
@@ -7,6 +7,8 @@ Parameters:
     Type: String
   Test4:
     Type: String
+  Test5:
+    Type: String
 
 Resources:
   Instance:
