--- tests/fixtures/nested_configs/sub/config.toml
@@ -1,8 +1,0 @@
-var_a = "cba"
-
-[env]
-AWS_PROFILE = "edited"
-AWS_DEFAULT_REGION = "eu-central-1"
-
-[sub.b]
-var_c = "sph"
